# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления автопарком.",
        "title": "Web-App API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/",
    "paths": {
        "/car": {
            "post": {
                "description": "Добавляет новый автомобиль",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "автомобиль"
                ],
                "summary": "Добавить новый автомобиль",
                "parameters": [
                    {
                        "description": "Автомобиль",
                        "name": "automobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddAutomobileDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            }
        },
        "/car/:id": {
            "get": {
                "description": "Возвращает  автомобиль",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "автомобиль"
                ],
                "summary": "Получить автомобиль по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID автомобиля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Получение автомобиля по id",
                        "name": "automobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Automobile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            },
            "delete": {
                "description": "Удаляет  автомобиль",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "автомобиль"
                ],
                "summary": "Удалить автомобиль",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID автомобиля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Удаление автомобиля",
                        "name": "automobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Automobile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            },
            "patch": {
                "description": "Обновляет  автомобиль",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "автомобиль"
                ],
                "summary": "Обновить автомобиль",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID автомобиля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные автомобиля",
                        "name": "automobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Automobile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            }
        },
        "/cars": {
            "get": {
                "description": "Возвращает  список автомобилей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "автомобиль"
                ],
                "summary": "Получить список автомобилей",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            }
        },
        "/inspection": {
            "post": {
                "description": "Добавляет тех осмотр",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "тех осмотр"
                ],
                "summary": "Добавить тех осмотр",
                "parameters": [
                    {
                        "description": "Техосмотр",
                        "name": "inspection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddInspectionDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            }
        },
        "/inspection/:id": {
            "get": {
                "description": "Возвращает тех осмотр",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "тех осмотр"
                ],
                "summary": "Получить тех осмотр по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тех осмотра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Получение тех осмотра по id",
                        "name": "automobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Inspection"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            },
            "delete": {
                "description": "Удаляет тех осмотр",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "тех осмотр"
                ],
                "summary": "Удалить тех осмотр по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тех осмотра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Удаление тех осмотра",
                        "name": "automobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Inspection"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            },
            "patch": {
                "description": "Обновляет  данные тех осмотра",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "тех осмотр"
                ],
                "summary": "Обновить данные о тех осмотре",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID техосмотра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные техосмотра",
                        "name": "inspection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateInspectionDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно"
                    },
                    "400": {
                        "description": "Неверные данные запроса"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            }
        },
        "/inspections": {
            "get": {
                "description": "Возвращает список тех осмотров",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "тех осмотр"
                ],
                "summary": "Получить список тех осмотров",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            }
        },
        "/report": {
            "get": {
                "description": "Возвращает агрегированные данные по автомобилям",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "автомобиль"
                ],
                "summary": "Получить отчёт по автомобилям",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Ошибка на сервере"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddAutomobileDTO": {
            "type": "object",
            "required": [
                "license_plate",
                "model",
                "registration_date",
                "release_date"
            ],
            "properties": {
                "license_plate": {
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 5
                },
                "model": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "registration_date": {
                    "$ref": "#/definitions/models.CustomDate"
                },
                "release_date": {
                    "$ref": "#/definitions/models.CustomDate"
                }
            }
        },
        "models.AddInspectionDTO": {
            "type": "object",
            "required": [
                "automobile_id",
                "card_number",
                "inspection_date"
            ],
            "properties": {
                "automobile_id": {
                    "type": "integer"
                },
                "card_number": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                },
                "inspection_date": {
                    "$ref": "#/definitions/models.CustomDate"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "models.Automobile": {
            "type": "object",
            "required": [
                "license_plate",
                "model",
                "registration_date",
                "release_date"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "license_plate": {
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 5
                },
                "model": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "registration_date": {
                    "$ref": "#/definitions/models.CustomDate"
                },
                "release_date": {
                    "$ref": "#/definitions/models.CustomDate"
                }
            }
        },
        "models.CustomDate": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "models.Inspection": {
            "type": "object",
            "required": [
                "automobile_id",
                "card_number",
                "inspection_date"
            ],
            "properties": {
                "automobile_id": {
                    "type": "integer"
                },
                "card_number": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                },
                "id": {
                    "type": "integer"
                },
                "inspection_date": {
                    "$ref": "#/definitions/models.CustomDate"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "models.UpdateInspectionDTO": {
            "type": "object",
            "required": [
                "automobile_id",
                "card_number",
                "inspection_date"
            ],
            "properties": {
                "automobile_id": {
                    "type": "integer"
                },
                "card_number": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                },
                "inspection_date": {
                    "$ref": "#/definitions/models.CustomDate"
                },
                "notes": {
                    "type": "string"
                }
            }
        }
    }
}